Welcome to Cards of Chaos, Game Design
This repository tracks the version history of Cards of Chaos, a tabletop roleplaying game with a new spin on combat encounters. Draft a deck of cards from the available classes and spell list to create a unique deck of your creation.  The cards you draw may be random, but your creativity creates the chaos.